Refactor the core module
This change focuses on moving code around to make `core` a little more
organized. It was getting out of hand.

- Create new `particle_map` module and move `ParticleType` out of
core::particle::mod.rs.
- Move a bunch of chunk-related tests out of core::particle, they must
have gotten left behind in a previous refactor
- Add new core::particle::registry module for the `ParticleTypeRegistry`
- Move ChanceLifetime and TimedLifetime to the lifecycle module.
core::mod is now purely for more basic type definitions
- Move ChanceMutation to the sync module
